From Gerhard Olsson:
Three parts. The second/third part is normally inactive in the code.

gtrnctr.c
Average Cadence in laps was handled as floats with one decimal but the xml format specifies char.
Some programs therefore failed to parse the generated xml
CadenceValue_t in http://www8.garmin.com/xmlschemas/TrainingCenterDatabasev2.xsd

jeeps
Cadence is invalid if it is 255 in Garmin USB. This is handled for track points but not for lap data.

garmin.c
(inactive) function read_lap_as_track() was not updated to the new string handling