mtk_logger mtk_parse_info: Fix track start for Holux devices
Previously, a new track was created when the event "GPS Logger turned on"
was read. This is correct for the MTK_LOGGER only, but not for the
Holux devices M241 & GR245.

Case #1: Firmware versions where the event is never logged. Effect: One large
track is created.
Applies to GR245 v1.x

Case #2: Firmware versions where the event is logged, but not
at the beginning of a track, but _after_ the first track point of a track
(at least most of the times).
Applies to GR245 v2.00, M241 v1.12, v1.13

For all of the currently known Holux firmware versions, the correct
indication of a track start is the "Log period change" /
"Log distance change" event.

This commit adds some more testcases (241 case provided by Erik Krause)
that were suffering from Case #2 before.

The existing 245 test is an example for Case #1 and has been fixed.