Fix default values for targetlocation and configlog flags.
The previous default value "" wereconsidered as TRUE in the IF
condition, Those flags took effect even if users didn't specify it.
There were no workarounds. Itprevented users to reach to the main
procedure skytraq_read_tracks.

Those should be NULL by default so that the IF condition should
be FALSE if a user wouldn't specify it.