Added support for FIT version 2.0 With FIT version 2.0 developer fields where added to the definition messages. These fields extends the definiton messages and due to this the number of records which must be parsed afterwards.
FIT files containing developer fields could not be parsed before
this changes.

Simple support was added by just parsing the data but without using them.
It is not expected that any of the currently used data like the
position or the elevation is coded inside the developer fields
because for this we have the already existing records from
FIT version 1.0

Added a test with a fit test file coming from the Wahoo Element Bolt
which uses developer fields from FIT version 2.0