Fix warning in .Pro build by forcing HAVE_UNISTD
I'm not sure if it's new in Qt 5.9, but it's trivial (I think) to fix and harmless. When building files like gzread and gzwrite, MacOS would complain that we were missing declarations to extremely well known system functions (write, close, read, seek, ...) Those live in <unistd.h>, so we nudge the Makefile generator to trigger its inclusion for MacOS and Linux.