clang tidy optimizes to use parameter passing by const reference.
RUN_CLANG_TIDY_FLAGS="-header-filter=.*\\\.h -checks=-*,performance-unnecessary-value-param -fix

The suggestion "parameter '*' is passed by value and only copied once; consider moving
it to avoid unnecessary copies" was not accepted.

The suggestion "the const qualified parameter '*' is copied for each invocation;
consider making it a reference" was accepted.

The suggestion "the parameter '*' is copied for each invocation but only used as a const
reference; consider making it a const reference" was accepted.