refactor track filter.
user visible changes:
Previously it was possible to cause a segmentation fault (due to a double delete)
if all the tracks were filtered out by the stop and start options.
This has been corrected.

Previously UTC time was used with the title option for split operations,
but local time was used with the title option for the pack operations.
Now UTC time is always used with the title option.

Previously the split and sdistance options allowed an option value
consisting of only the units.  This was undocumented.  Now both a number
and the units are required.

Previously the split option documentation stated that units of seconds
would be used if none were specified, however the units were actually required.
The documenation is updated to reflect historical and current usage.

Previously the split and sdistance options allowed the units to precede or
follow the quantity.  The documentation stated the units must follow the
quantity.  Now the units must follow the quantity.

Previously synthesized speeds ignored the fractional part of the track point times,
now the fractional part of the track point times are used in the speed calculation.

refactoring changes:
time_t is replaced where possible by gpsbabel::DateTime or QDateTime.

option value parsing is done with QRegularExpression.

some integer types were converted to bool types.

test cases are expanded for better coverage.