fiddle with nmea waypoint status tracking.
Traditionally nmea looks at the Waypoint data member Q
to decide if the Waypoint has been added on to a global list, and
if it has not it deletes the Waypoint so we don't leak memory.

This is problematic when we change the waypoint lists so queues
aren't used.

To overcome this difficulty nmea now uses a derived class of
Waypoint that tracks if the Waypoint has been added or not.