harden gdb reader to long string input.
fix bug in the gdb reader that resulted in file position being lost
if a string was truncated when reading due to the finite buffer size.

fix bug in the gdb reader that could result in data being written outside
the provided buffer.

enhance gdb readers to remove length limit on reading strings.