Eliminate some magic numbers for unit conversion.
Previously the conversion factor for meters to feet was inaccurate.  Correcting
this caused a couple kml test cases to change.

The ozi test was problematic, an altitiude was provided that endend in
0.5 feet, and was printed out rounded to feet.  A one bit error in the
mantissa of the double involved in the round trip to meters could change
the rounding result.  To fix this the ozi writer was changed to print
track altitudes to a tenth of a foot.  This matches an ozi track file found in
the wild as well as our own ozi track sample.

The classic-2 test involving gnuplot.style printed altitude to one millonth
of a foot, i.e. 0.3 micrometers!  The gnu plot style was changed to print
to the thousandth of a foot, i.e. 0.3 millimeters.