fix gtm writer issue with empty tracks.
This format expects one track style entry for each tracklog entry
with the tracklog flag set.  If the tracklog flag is set it indicates the
start of a new track.  Each tracklog entry represents a point.
Thus, it is not possible to represent a track without any points in this
format.

This change prevents the writing of track style entries for tracks that
don't have any points.

Previously, this could cause counts from test-all in the category
"tests without error but with unexpected output" if the refernce file
had tracks with no points.