update linux/macos CI Qt 5.12.8 (#541)
* update CI to 5.12.8 (except appveyor)

we use the appveyor installed Qt 5.12, which is currently
5.12.6.

* update travis yaml syntax.

* update macos build system for Qt 5.12 build.

* try osx_image: xcode11.3

osx_image xcode10.3 failed 3 times in a row.  It showed an
incomplete Qt install and repeated attempts to detatch.

* debug macos qt install fails

* Revert "debug macos qt install fails"

This reverts commit 36df50bc865b515b4545f5f25cc65f5e0575d325.

* Revert "try osx_image: xcode11.3"

This reverts commit ff1e803234b31671e77d8d9ca68e74d62e753b80.

* Revert "update macos build system for Qt 5.12 build."

This reverts commit efedf5132a84141ae351cfe308433bfbda53374b.

* cleanup new docker script.

* build docker image in clean directory.

* move macos Qt 5.12 build to xcode 10.3.

This includes a workaround for a bug in the Qt intaller
that causes segmentation faults in the Qt installer when
the minimal platform is used on macos 10.14, 10.15.