convert unicsv to Format class.
Two bugs are fixed as well:
1. { "datum",  fld_date, STR_ANY } is removed from the fields_def.
Note that the type is fld_date.  I don't think this was ever used.
We do have the datum option to set the datum from the command line.
2. unicsv_fondle_header changed the fields_def type to fld_iso_time
if it was fld_time or fld_date and the value contained "iso".
However, fields_def isn't used outside unicsv_fondle_header, so
this had no effect.