fix nmea writer date bug. (#553)
* fix nmea writer date bug.

dates before 2000 were written with a month off by one.

This also changes the behavior when the creation time is invalid.
Previously you would get data related to the unix epoch for the date
and time fields.
Now you get empty fields, i.e. ",,"

* fix magellan writer date bug.

This had the same bug as nmea.  It had an additional bug
when it attempted to round the fractional part of a second.  Rounding
can ripple all the way from the fractional part of a second to the year,
as is now demonstrated in the testcase.

Like nmea, magellan will now print empty fields if the creation time
is invalid.

* fix date bug in gpssim.

This is the same bug as in nmea, magproto.