fix -Wstringop-truncation warning in enigma.
These are nonstrings.  They are not required to be null terminated.

A real reference file caputred from MGL Central 2.0 is added.  In it
you can see that don't care values are not all filled with a particular
value.  We fill them with NULLs are a result of zeroing the structure before
we write it.  You can also see the 1000m offset in altitudes is correct.