support doc creation in qmake.
pass parameters to tools/mkcapabilites instead of
using text replacement in configure.

text replacement in xmldoc/makedoc was unnecessary, directory
creation is handled in the recipee.

there is some foolishness in GPSBabel.pro to allow overriding of
WEB and DOCVERSION on the command line of make.  If those values
are known when running qmake it could be slightly simpler.