use scripts for document creation.
this simplifies error handling and quoting.

qmake creates a makefile with all the commands for a QMAKE_EXTRA_TARGET
as one line in a recipe.  Make passes that line to the shell with a -c.
As a result any errors in any of the commands in the series of commands
except for the last will not be seen by make.

fix shellcheck issues with fixdoc, mkcapabilities scripts.