correct hiketech written time values.
the mystery previously noted in the code was due to the use of the %I format
specifier which prints out the hour using a 12 hour clock.

note that it is easy to see the times in the reference files were wrong.  the
times in the first reference file, expertgpx.gpx, are in UTC.  This is always
true for gpx, and is indicated as well by the Z.  the times in the hiketech file
appear in a utc element, so they must be in UTC as well.

this eliminates the usage of QDateTime::toTime_t which is obsolete.