Move destinator from ctime to QDateTime.
This feels a bit like rearranging Titanic chairs, but I needed to reduce
callers of strptime. This is another format I think we've rewritten a
couple of times without evidence of users. Most of the traffic in the last
ten years has been complaints of it not working

Destinator, consider yourself warned. You're on the death watch...