use db5-upgrade.xsl from docbook 5.1 release.
we are still converting to docbook 5.0.

this fixes issues with missing Appendix labels
(e.g. "Appendix A. Supported Datums" was "Suppored Datums")
and Glossary Indexes (e.g. "G" was omitted)
in the html output.

unfortunately it causes issues with translating ulinks inside term
elements.  We patch the dropped links back in.