update documentation to DocBook 5.0 (#940)
* convert to docbook 5.0.

* get jing running.

* use uri to find docbook rng.

* brew install jing-trang.  use oasis for rngs.

* add insertion markers to autogen doc files.

* use db5-upgrade.xsl from docbook 5.1 release.

we are still converting to docbook 5.0.

this fixes issues with missing Appendix labels
(e.g. "Appendix A. Supported Datums" was "Suppored Datums")
and Glossary Indexes (e.g. "G" was omitted)
in the html output.

unfortunately it causes issues with translating ulinks inside term
elements.  We patch the dropped links back in.

* make sure makedoc is using utf8 for files.