wrap long lines in verbatim elements, such as screen.
The use of inline.boldmonoseq on verbatim elements such as screen
results in a lose of the line endings. This makes it impossible to
differentiate items in what might have better been a list.  However,
if the line in the screen element is too long it will be trucated.
Use the monospace.verbatim.property wrap-option to wrap any such lines.

An example that illustrates the problems is the list of keywords in the
unicsv format.