don't create fictious course and speed values.
We compute coures and speed in track_recompute, which is used
by the kml and gtrnctr writers.  We also compute these in the track
filter with the course speed options.

Both of these algorithums compute the course and speed for a point
based on the time and position of that point and a previous point.
Given that no information from the future is used, the course and
speed of the first point cannot be determined.  This PR introduces
a change so a fictious value for the first point is not generated.