add testing for kml:dateTimeType.
and fix a related bug.

Note that we incorrectly, for kml:dateTimeType, force values
without time zone qualifiers to UTC.
We do not test the optional gYearMonth time zone qualifer.
We do not test the optional gYear time zone qualifer.
We do not test the optional date time zone qualifer.