resolve maintanence issue in interpolate & resample filters. (#1161)
* resolve maintanence issue in interpolate filter.

Instead of manually creating a deep copy of the route list with
an empty waypoint list we operate on the original route list by
swapping it's waypoint list with an empty list.  This is both more
efficient and easier to maintain.

* resolve maintanence issues in resample filter.

Instead of manually creating a deep copy of the route list with
an empty waypoint list we operate on the original route list by
swapping it's waypoint list with an empty list.  This is both more
efficient and easier to maintain.

For efficient deletion use track_del_marked_wpts.