fix potentially wrong request for size to malloc in zlib
item and size are multiplied, but are potentially signed. If an overflow turns into an underflow, this malloc/calloc are likely doomed.

This comes from #47. 

I don't like this patch and may not submit this at all depending on what upstream does. This is mostly to see how presubmit like it.