fix bugs with gpx passthrough.
with gpx 1.0, any extension data that was a child of rte or trk
was lost.  I beleive this used to work but has been broken for quite
a while.

with gpx 1.0 extension data that was a child of gpx could be
erroneously rewritten as part of the final wpt/rtept/trkpt.

with gpx 1.1 extension data that was a chile of trkseg could be
erroneously rewritten as part of the final wpt/rtept/trkpt.

I note we currently drop any extension data that is a child of gpx,
,a child of trkseg, or a child of metadata.