support optional start elements in xmlstreamwriter. (#1193)
this adds methods to xmlstreamwriter to allow conditional writing
of a start element.  The conditional element will only be written to
the stream if it has children, which is deteremined when the condtional
element is ended.  This simplifies writing of such elements as the
writer does not need to know if it will subsequently write children or
not.