jeeps function pointers, GPS_PPacket const correctness, GPS_Serial_[OSP]Packet -> GPS_Serial_Packet (#1215)
* use using for jeeps function prototypes.

* const correctness wrt GPS_PPacket.

The important bit here is in gpssend.cc, Build_Serial_Packet used to rececive a 1KB packet by copy, now it is by const reference.

* delete accidentally committed file.

* rename GPS_PPacket -> GPS_Packet.

original jeeps used GPS_SXXX, GPS_OSXXX, GPS_PXXX for the structure
tag, a typedef of the structure, and a pointer to a typedef of the
structure.  In the case of GPS_PPacket this convention was no longer
followed.

* change GPS_Packet back to a structure.

* try harder to change GPS_Packet back to a struct.

* clean up GPS_Serial_[SOP]Packet.

* catch some missed renames up.

clion on linux missed some windows only renames.