correct WGS84 semi minor axis value.
It appears the GSR80 value was used instead of the WGS84 value, which
is a very small error.
However, there was a typo in the value in GPS_Math_XYZ_To_WGS84LatLonH
that resulted in an order of magnitude error!
This could lead to GPS_Math_XYZ_To_WGS84LatLonH never converging.
This may be why skytraq created their own function to convert ECEF
to WGS84.