Add Kalman doc.
Try to make Kalman understandable to our audience. Shows how it's
applied, with visual (including animations for the web!) presentation
when it helps.

The PDF looks great, IMO.
The HTML generation probably needs help and i don't know if
a 900K GIF will fly, but IMO it _really_ helps.

I'm pretty proud of all this, too. :-)